Rename binder transport example client app
Rename the app to exampleclient

Also moved and updated README.md
