+++ examples/android/binder/java/io/grpc/binder/cpp/exampleclient/res/layout/activity_main.xml
