+++ examples/android/binder/java/io/grpc/binder/cpp/exampleclient/res/values/strings.xml
